// frontend/src/app/generated/event.ts
/**
 * RESTful API of iHRM
 * The RESTful API of the main core of the iHRM application. Add-ons will have their own API documentation.
 *
 * OpenAPI spec version: 2020.06.14
 * Contact: ihrm@imbus.tn
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Event { 
    id: number;
    calendarId: number;
    isPrivateCalendar: boolean;
    start: string;
    end: string;
    title: string;
    description: string;
    repetition: Event.RepetitionEnum;
    allDay: boolean;
    eventType: Event.EventTypeEnum;
    creator: number;
}
export namespace Event {
    export type RepetitionEnum = 'Daily' | 'Weekly' | 'Monthly' | 'Yearly' | 'Unrepeatable';
    export const RepetitionEnum = {
        Daily: 'Daily' as RepetitionEnum,
        Weekly: 'Weekly' as RepetitionEnum,
        Monthly: 'Monthly' as RepetitionEnum,
        Yearly: 'Yearly' as RepetitionEnum,
        Unrepeatable: 'Unrepeatable' as RepetitionEnum
    };
    export type EventTypeEnum = 'Meeting' | 'Workshop' | 'Training' | 'Leave';
    export const EventTypeEnum = {
        Meeting: 'Meeting' as EventTypeEnum,
        Workshop: 'Workshop' as EventTypeEnum,
        Training: 'Training' as EventTypeEnum,
        Leave: 'Leave' as EventTypeEnum
    };
}